+++ packages/ui/src/hooks/useSessionStorage.ts
@@ -1,5 +1,3 @@
-'use client';
-
 import { type Dispatch, type SetStateAction, useCallback, useEffect, useState } from 'react';
 
 import { useEventCallback } from './useEventCallback';